+++ store/src/Haskoin/Store/Database/Writer.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DeriveAnyClass    #-}
+{-# LANGUAGE DeriveGeneric     #-}
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE LambdaCase        #-}
@@ -8,72 +10,136 @@ module Haskoin.Store.Database.Writer
     ) where
 
 import           Control.Applicative           ((<|>))
-import           Control.Monad                 (join)
+import           Control.DeepSeq               (NFData)
 import           Control.Monad.Except          (MonadError)
 import           Control.Monad.Reader          (ReaderT)
 import qualified Control.Monad.Reader          as R
 import           Control.Monad.Trans.Maybe     (MaybeT (..), runMaybeT)
+import qualified Data.ByteString.Short         as B.Short
+import           Data.Hashable                 (Hashable)
 import           Data.HashMap.Strict           (HashMap)
 import qualified Data.HashMap.Strict           as M
 import           Data.IntMap.Strict            (IntMap)
 import qualified Data.IntMap.Strict            as I
-import           Data.List                     (nub)
-import           Data.Maybe                    (fromJust, fromMaybe, isJust,
-                                                maybeToList)
+import           Data.Maybe                    (fromJust, isJust, maybeToList)
 import           Database.RocksDB              (BatchOp)
 import           Database.RocksDB.Query        (deleteOp, insertOp, writeBatch)
+import           GHC.Generics                  (Generic)
 import           Haskoin                       (Address, BlockHash, BlockHeight,
-                                                OutPoint (..), TxHash)
+                                                OutPoint (..), TxHash,
+                                                headerHash, txHash)
 import           Haskoin.Store.Common          (StoreRead (..), StoreWrite (..))
-import           Haskoin.Store.Data            (Balance, BlockData,
-                                                BlockRef (..), TxRef (..),
-                                                Spender, TxData, Unspent,
-                                                nullBalance, zeroBalance)
-import           Haskoin.Store.Database.Memory (MemoryDatabase (..),
-                                                MemoryState (..),
-                                                emptyMemoryDatabase,
-                                                getMempoolH, getSpenderH,
-                                                getSpendersH, getUnspentH,
-                                                withMemoryDatabase)
+import           Haskoin.Store.Data            (Balance (..), BlockData (..),
+                                                BlockRef (..), Spender,
+                                                TxData (..), TxRef (..),
+                                                Unspent (..))
 import           Haskoin.Store.Database.Reader (DatabaseReader (..),
                                                 withDatabaseReader)
 import           Haskoin.Store.Database.Types  (AddrOutKey (..), AddrTxKey (..),
                                                 BalKey (..), BalVal (..),
                                                 BestKey (..), BlockKey (..),
                                                 HeightKey (..), MemKey (..),
-                                                OutVal, SpenderKey (..),
+                                                OutVal (..), SpenderKey (..),
                                                 TxKey (..), UnspentKey (..),
-                                                UnspentVal (..))
-import           UnliftIO                      (MonadIO, newTVarIO, readTVarIO)
+                                                UnspentVal (..), balanceToVal,
+                                                unspentToVal, valToBalance,
+                                                valToUnspent)
+import           UnliftIO                      (MonadIO, TVar, atomically,
+                                                modifyTVar, newTVarIO,
+                                                readTVarIO)
+
+data Dirty a = Modified a | Deleted
+    deriving (Eq, Show, Generic, Hashable, NFData)
+
+instance Functor Dirty where
+    fmap _ Deleted      = Deleted
+    fmap f (Modified a) = Modified (f a)
 
 data DatabaseWriter = DatabaseWriter
     { databaseWriterReader :: !DatabaseReader
-    , databaseWriterState  :: !MemoryState
+    , databaseWriterState  :: !(TVar MemoryDatabase)
     }
+
+data MemoryDatabase = MemoryDatabase
+    { hBest
+      :: !(Maybe BlockHash)
+    , hBlock
+      :: !(HashMap BlockHash BlockData)
+    , hHeight
+      :: !(HashMap BlockHeight [BlockHash])
+    , hTx
+      :: !(HashMap TxHash TxData)
+    , hSpender
+      :: !(HashMap TxHash (IntMap (Dirty Spender)))
+    , hUnspent
+      :: !(HashMap TxHash (IntMap (Dirty UnspentVal)))
+    , hBalance
+      :: !(HashMap Address BalVal)
+    , hAddrTx
+      :: !(HashMap Address (HashMap TxRef (Dirty ())))
+    , hAddrOut
+      :: !(HashMap Address (HashMap BlockRef (HashMap OutPoint (Dirty OutVal))))
+    , hMempool
+      :: !(Maybe [TxRef])
+    } deriving (Eq, Show)
+
+instance MonadIO m => StoreWrite (ReaderT (TVar MemoryDatabase) m) where
+    setBest h = do
+        v <- R.ask
+        atomically $ modifyTVar v (setBestH h)
+    insertBlock b = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertBlockH b)
+    setBlocksAtHeight h g = do
+        v <- R.ask
+        atomically $ modifyTVar v (setBlocksAtHeightH h g)
+    insertTx t = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertTxH t)
+    insertSpender p s = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertSpenderH p s)
+    deleteSpender p = do
+        v <- R.ask
+        atomically $ modifyTVar v (deleteSpenderH p)
+    insertAddrTx a t = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertAddrTxH a t)
+    deleteAddrTx a t = do
+        v <- R.ask
+        atomically $ modifyTVar v (deleteAddrTxH a t)
+    insertAddrUnspent a u = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertAddrUnspentH a u)
+    deleteAddrUnspent a u = do
+        v <- R.ask
+        atomically $ modifyTVar v (deleteAddrUnspentH a u)
+    setMempool xs = do
+        v <- R.ask
+        atomically $ modifyTVar v (setMempoolH xs)
+    setBalance b = do
+        v <- R.ask
+        atomically $ modifyTVar v (setBalanceH b)
+    insertUnspent h = do
+        v <- R.ask
+        atomically $ modifyTVar v (insertUnspentH h)
+    deleteUnspent p = do
+        v <- R.ask
+        atomically $ modifyTVar v (deleteUnspentH p)
 
 runDatabaseWriter ::
        (MonadIO m, MonadError e m)
     => DatabaseReader
     -> ReaderT DatabaseWriter m a
     -> m a
-runDatabaseWriter bdb@DatabaseReader { databaseHandle = db
-                                     , databaseMaxGap = gap
-                                     , databaseInitialGap = igap
-                                     , databaseNetwork = net
-                                     } f = do
+runDatabaseWriter bdb@DatabaseReader{databaseHandle = db} f = do
     hm <- newTVarIO emptyMemoryDatabase
-    let ms =
-            MemoryState
-                { memoryDatabase = hm
-                , memoryMaxGap = gap
-                , memoryNetwork = net
-                , memoryInitialGap = igap
-                }
     x <-
         R.runReaderT
             f
             DatabaseWriter
-                {databaseWriterReader = bdb, databaseWriterState = ms}
+                { databaseWriterReader = bdb
+                , databaseWriterState = hm }
     ops <- hashMapOps <$> readTVarIO hm
     writeBatch db ops
     return x
@@ -110,12 +176,12 @@ txOps = map (uncurry f) . M.toList
   where
     f = insertOp . TxKey
 
-spenderOps :: HashMap TxHash (IntMap (Maybe Spender)) -> [BatchOp]
+spenderOps :: HashMap TxHash (IntMap (Dirty Spender)) -> [BatchOp]
 spenderOps = concatMap (uncurry f) . M.toList
   where
     f h = map (uncurry (g h)) . I.toList
-    g h i (Just s) = insertOp (SpenderKey (OutPoint h (fromIntegral i))) s
-    g h i Nothing  = deleteOp (SpenderKey (OutPoint h (fromIntegral i)))
+    g h i (Modified s) = insertOp (SpenderKey (OutPoint h (fromIntegral i))) s
+    g h i Deleted      = deleteOp (SpenderKey (OutPoint h (fromIntegral i)))
 
 balOps :: HashMap Address BalVal -> [BatchOp]
 balOps = map (uncurry f) . M.toList
@@ -123,45 +189,25 @@ balOps = map (uncurry f) . M.toList
     f = insertOp . BalKey
 
 addrTxOps ::
-       HashMap Address (HashMap BlockRef (HashMap TxHash Bool)) -> [BatchOp]
-addrTxOps = concat . concatMap (uncurry f) . M.toList
+       HashMap Address (HashMap TxRef (Dirty ())) -> [BatchOp]
+addrTxOps = concatMap (uncurry f) . M.toList
   where
     f a = map (uncurry (g a)) . M.toList
-    g a b = map (uncurry (h a b)) . M.toList
-    h a b t True =
-        insertOp
-            (AddrTxKey
-                 { addrTxKeyA = a
-                 , addrTxKeyT =
-                       TxRef
-                           { txRefBlock = b
-                           , txRefHash = t
-                           }
-                 })
-            ()
-    h a b t False =
-        deleteOp
-            AddrTxKey
-                { addrTxKeyA = a
-                , addrTxKeyT =
-                      TxRef
-                          { txRefBlock = b
-                          , txRefHash = t
-                          }
-                }
+    g a t (Modified ()) = insertOp (AddrTxKey a t) ()
+    g a t Deleted       = deleteOp (AddrTxKey a t)
 
 addrOutOps ::
-       HashMap Address (HashMap BlockRef (HashMap OutPoint (Maybe OutVal)))
+       HashMap Address (HashMap BlockRef (HashMap OutPoint (Dirty OutVal)))
     -> [BatchOp]
 addrOutOps = concat . concatMap (uncurry f) . M.toList
   where
     f a = map (uncurry (g a)) . M.toList
     g a b = map (uncurry (h a b)) . M.toList
-    h a b p (Just l) =
+    h a b p (Modified l) =
         insertOp
             (AddrOutKey {addrOutKeyA = a, addrOutKeyB = b, addrOutKeyP = p})
             l
-    h a b p Nothing =
+    h a b p Deleted =
         deleteOp AddrOutKey {addrOutKeyA = a, addrOutKeyB = b, addrOutKeyP = p}
 
 mempoolOp :: [TxRef] -> BatchOp
@@ -169,12 +215,12 @@ mempoolOp =
     insertOp MemKey .
     map (\TxRef {txRefBlock = MemRef t, txRefHash = h} -> (t, h))
 
-unspentOps :: HashMap TxHash (IntMap (Maybe UnspentVal)) -> [BatchOp]
+unspentOps :: HashMap TxHash (IntMap (Dirty UnspentVal)) -> [BatchOp]
 unspentOps = concatMap (uncurry f) . M.toList
   where
     f h = map (uncurry (g h)) . I.toList
-    g h i (Just u) = insertOp (UnspentKey (OutPoint h (fromIntegral i))) u
-    g h i Nothing  = deleteOp (UnspentKey (OutPoint h (fromIntegral i)))
+    g h i (Modified u) = insertOp (UnspentKey (OutPoint h (fromIntegral i))) u
+    g h i Deleted      = deleteOp (UnspentKey (OutPoint h (fromIntegral i)))
 
 setBestI :: MonadIO m => BlockHash -> DatabaseWriter -> m ()
 setBestI bh DatabaseWriter {databaseWriterState = hm} =
@@ -184,7 +230,8 @@ insertBlockI :: MonadIO m => BlockData -> DatabaseWriter -> m ()
 insertBlockI b DatabaseWriter {databaseWriterState = hm} =
     withMemoryDatabase hm $ insertBlock b
 
-setBlocksAtHeightI :: MonadIO m => [BlockHash] -> BlockHeight -> DatabaseWriter -> m ()
+setBlocksAtHeightI
+    :: MonadIO m => [BlockHash] -> BlockHeight -> DatabaseWriter -> m ()
 setBlocksAtHeightI hs g DatabaseWriter {databaseWriterState = hm} =
     withMemoryDatabase hm $ setBlocksAtHeight hs g
 
@@ -217,81 +264,86 @@ deleteAddrUnspentI a u DatabaseWriter {databaseWriterState = hm} =
     withMemoryDatabase hm $ deleteAddrUnspent a u
 
 setMempoolI :: MonadIO m => [TxRef] -> DatabaseWriter -> m ()
-setMempoolI xs DatabaseWriter {databaseWriterState = hm} = withMemoryDatabase hm $ setMempool xs
+setMempoolI xs DatabaseWriter {databaseWriterState = hm} =
+    withMemoryDatabase hm $ setMempool xs
 
 getBestBlockI :: MonadIO m => DatabaseWriter -> m (Maybe BlockHash)
-getBestBlockI DatabaseWriter {databaseWriterState = hm, databaseWriterReader = db} =
+getBestBlockI
+    DatabaseWriter {databaseWriterState = hm, databaseWriterReader = db} =
     runMaybeT $ MaybeT f <|> MaybeT g
   where
-    f = withMemoryDatabase hm getBestBlock
+    f = getBestBlockH <$> readTVarIO hm
     g = withDatabaseReader db getBestBlock
 
 getBlocksAtHeightI :: MonadIO m => BlockHeight -> DatabaseWriter -> m [BlockHash]
-getBlocksAtHeightI bh DatabaseWriter {databaseWriterState = hm, databaseWriterReader = db} = do
-    xs <- withMemoryDatabase hm $ getBlocksAtHeight bh
-    ys <- withDatabaseReader db $ getBlocksAtHeight bh
-    return . nub $ xs <> ys
+getBlocksAtHeightI
+    bh
+    DatabaseWriter {databaseWriterState = hm, databaseWriterReader = db} =
+    getBlocksAtHeightH bh <$> readTVarIO hm >>= \case
+    Just bs -> return bs
+    Nothing -> withDatabaseReader db $ getBlocksAtHeight bh
 
 getBlockI :: MonadIO m => BlockHash -> DatabaseWriter -> m (Maybe BlockData)
-getBlockI bh DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
+getBlockI
+    bh
+    DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
     runMaybeT $ MaybeT f <|> MaybeT g
   where
-    f = withMemoryDatabase hm $ getBlock bh
+    f = getBlockH bh <$> readTVarIO hm
     g = withDatabaseReader db $ getBlock bh
 
 getTxDataI ::
        MonadIO m => TxHash -> DatabaseWriter -> m (Maybe TxData)
-getTxDataI th DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
+getTxDataI
+    th
+    DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
     runMaybeT $ MaybeT f <|> MaybeT g
   where
-    f = withMemoryDatabase hm $ getTxData th
+    f = getTxDataH th <$> readTVarIO hm
     g = withDatabaseReader db $ getTxData th
 
 getSpenderI :: MonadIO m => OutPoint -> DatabaseWriter -> m (Maybe Spender)
-getSpenderI op DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
-    fmap join . runMaybeT $ MaybeT f <|> MaybeT g
-  where
-    f = getSpenderH op <$> readTVarIO (memoryDatabase hm)
-    g = Just <$> withDatabaseReader db (getSpender op)
+getSpenderI
+    op
+    DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} =
+    getSpenderH op <$> readTVarIO hm >>= \case
+    Just (Modified s) -> return (Just s)
+    Just Deleted -> return Nothing
+    Nothing -> withDatabaseReader db (getSpender op)
 
 getSpendersI :: MonadIO m => TxHash -> DatabaseWriter -> m (IntMap Spender)
-getSpendersI t DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} = do
-    hsm <- getSpendersH t <$> readTVarIO (memoryDatabase hm)
-    dsm <- I.map Just <$> withDatabaseReader db (getSpenders t)
-    return . I.map fromJust . I.filter isJust $ hsm <> dsm
+getSpendersI
+    t
+    DatabaseWriter {databaseWriterReader = db, databaseWriterState = hm} = do
+    hsm <- fmap to_maybe . getSpendersH t <$> readTVarIO hm
+    dsm <- fmap Just <$> withDatabaseReader db (getSpenders t)
+    return $ I.map fromJust $ I.filter isJust $ hsm <> dsm
+  where
+    to_maybe Deleted      = Nothing
+    to_maybe (Modified x) = Just x
 
 getBalanceI :: MonadIO m => Address -> DatabaseWriter -> m Balance
-getBalanceI a DatabaseWriter { databaseWriterReader = db
-                             , databaseWriterState = hm
-                             } =
-    fromMaybe (zeroBalance a) <$> runMaybeT (MaybeT f <|> MaybeT g)
-  where
-    f =
-        withMemoryDatabase hm $
-        getBalance a >>= \b ->
-            return $
-            if nullBalance b
-                then Nothing
-                else Just b
-    g =
-        withDatabaseReader db $
-        getBalance a >>= \b ->
-            return $
-            if nullBalance b
-                then Nothing
-                else Just b
+getBalanceI
+    a
+    DatabaseWriter { databaseWriterReader = db, databaseWriterState = hm} =
+    getBalanceH a <$> readTVarIO hm >>= \case
+    Just b -> return $ valToBalance a b
+    Nothing -> withDatabaseReader db $ getBalance a
 
 setBalanceI :: MonadIO m => Balance -> DatabaseWriter -> m ()
 setBalanceI b DatabaseWriter {databaseWriterState = hm} =
     withMemoryDatabase hm $ setBalance b
 
 getUnspentI :: MonadIO m => OutPoint -> DatabaseWriter -> m (Maybe Unspent)
-getUnspentI op DatabaseWriter { databaseWriterReader = db
-                              , databaseWriterState = hm
-                              } = fmap join . runMaybeT $ MaybeT f <|> MaybeT g
-  where
-    f = getUnspentH op <$> readTVarIO (memoryDatabase hm)
-    g = Just <$> withDatabaseReader db (getUnspent op)
+getUnspentI
+    op
+    DatabaseWriter
+    { databaseWriterReader = db
+    , databaseWriterState = hm
+    } = getUnspentH op <$> readTVarIO hm >>= \case
+    Just Deleted -> return Nothing
+    Just (Modified u) -> return (Just (valToUnspent op u))
+    Nothing -> withDatabaseReader db (getUnspent op)
 
 insertUnspentI :: MonadIO m => Unspent -> DatabaseWriter -> m ()
 insertUnspentI u DatabaseWriter {databaseWriterState = hm} =
@@ -306,7 +358,7 @@ getMempoolI ::
     => DatabaseWriter
     -> m [TxRef]
 getMempoolI DatabaseWriter {databaseWriterState = hm, databaseWriterReader = db} =
-    getMempoolH <$> readTVarIO (memoryDatabase hm) >>= \case
+    getMempoolH <$> readTVarIO hm >>= \case
         Just xs -> return xs
         Nothing -> withDatabaseReader db getMempool
 
@@ -341,3 +393,172 @@ instance MonadIO m => StoreWrite (ReaderT DatabaseWriter m) where
     insertUnspent u = R.ask >>= insertUnspentI u
     deleteUnspent p = R.ask >>= deleteUnspentI p
     setBalance b = R.ask >>= setBalanceI b
+
+withMemoryDatabase ::
+       MonadIO m
+    => TVar MemoryDatabase
+    -> ReaderT (TVar MemoryDatabase) m a
+    -> m a
+withMemoryDatabase = flip R.runReaderT
+
+emptyMemoryDatabase :: MemoryDatabase
+emptyMemoryDatabase =
+    MemoryDatabase
+        { hBest = Nothing
+        , hBlock = M.empty
+        , hHeight = M.empty
+        , hTx = M.empty
+        , hSpender = M.empty
+        , hUnspent = M.empty
+        , hBalance = M.empty
+        , hAddrTx = M.empty
+        , hAddrOut = M.empty
+        , hMempool = Nothing
+        }
+
+getBestBlockH :: MemoryDatabase -> Maybe BlockHash
+getBestBlockH = hBest
+
+getBlocksAtHeightH :: BlockHeight -> MemoryDatabase -> Maybe [BlockHash]
+getBlocksAtHeightH h = M.lookup h . hHeight
+
+getBlockH :: BlockHash -> MemoryDatabase -> Maybe BlockData
+getBlockH h = M.lookup h . hBlock
+
+getTxDataH :: TxHash -> MemoryDatabase -> Maybe TxData
+getTxDataH t = M.lookup t . hTx
+
+getSpenderH :: OutPoint -> MemoryDatabase -> Maybe (Dirty Spender)
+getSpenderH op db = do
+    m <- M.lookup (outPointHash op) (hSpender db)
+    I.lookup (fromIntegral (outPointIndex op)) m
+
+getSpendersH :: TxHash -> MemoryDatabase -> IntMap (Dirty Spender)
+getSpendersH t = M.lookupDefault I.empty t . hSpender
+
+getBalanceH :: Address -> MemoryDatabase -> Maybe BalVal
+getBalanceH a = M.lookup a . hBalance
+
+getMempoolH :: MemoryDatabase -> Maybe [TxRef]
+getMempoolH = hMempool
+
+setBestH :: BlockHash -> MemoryDatabase -> MemoryDatabase
+setBestH h db = db {hBest = Just h}
+
+insertBlockH :: BlockData -> MemoryDatabase -> MemoryDatabase
+insertBlockH bd db =
+    db { hBlock = M.insert (headerHash (blockDataHeader bd)) bd (hBlock db) }
+
+setBlocksAtHeightH
+    :: [BlockHash] -> BlockHeight -> MemoryDatabase -> MemoryDatabase
+setBlocksAtHeightH hs g db = db {hHeight = M.insert g hs (hHeight db)}
+
+insertTxH :: TxData -> MemoryDatabase -> MemoryDatabase
+insertTxH tx db = db {hTx = M.insert (txHash (txData tx)) tx (hTx db)}
+
+insertSpenderH :: OutPoint -> Spender -> MemoryDatabase -> MemoryDatabase
+insertSpenderH op s db =
+    db
+        { hSpender =
+              M.insertWith
+                  (<>)
+                  (outPointHash op)
+                  (I.singleton (fromIntegral (outPointIndex op)) (Modified s))
+                  (hSpender db)
+        }
+
+deleteSpenderH :: OutPoint -> MemoryDatabase -> MemoryDatabase
+deleteSpenderH op db =
+    db
+        { hSpender =
+              M.insertWith
+                  (<>)
+                  (outPointHash op)
+                  (I.singleton (fromIntegral (outPointIndex op)) Deleted)
+                  (hSpender db)
+        }
+
+setBalanceH :: Balance -> MemoryDatabase -> MemoryDatabase
+setBalanceH bal db =
+    db {hBalance = M.insert (balanceAddress bal) b (hBalance db)}
+  where
+    b = balanceToVal bal
+
+insertAddrTxH :: Address -> TxRef -> MemoryDatabase -> MemoryDatabase
+insertAddrTxH a btx db =
+    db
+        { hAddrTx =
+              M.insertWith
+                  (<>)
+                  a
+                  (M.singleton btx (Modified ()))
+                  (hAddrTx db)
+        }
+
+deleteAddrTxH :: Address -> TxRef -> MemoryDatabase -> MemoryDatabase
+deleteAddrTxH a btx db =
+    db
+        { hAddrTx =
+              M.insertWith
+                  (<>)
+                  a
+                  (M.singleton btx Deleted)
+                  (hAddrTx db)
+        }
+
+insertAddrUnspentH :: Address -> Unspent -> MemoryDatabase -> MemoryDatabase
+insertAddrUnspentH a u db =
+    let uns =
+            OutVal
+                { outValAmount = unspentAmount u
+                , outValScript = B.Short.fromShort (unspentScript u)
+                }
+        s =
+            M.singleton
+                a
+                (M.singleton
+                     (unspentBlock u)
+                     (M.singleton (unspentPoint u) (Modified uns)))
+     in db {hAddrOut = M.unionWith (M.unionWith M.union) s (hAddrOut db)}
+
+deleteAddrUnspentH :: Address -> Unspent -> MemoryDatabase -> MemoryDatabase
+deleteAddrUnspentH a u db =
+    let s =
+            M.singleton
+                a
+                (M.singleton
+                     (unspentBlock u)
+                     (M.singleton (unspentPoint u) Deleted))
+     in db {hAddrOut = M.unionWith (M.unionWith M.union) s (hAddrOut db)}
+
+setMempoolH :: [TxRef] -> MemoryDatabase -> MemoryDatabase
+setMempoolH xs db = db {hMempool = Just xs}
+
+getUnspentH :: OutPoint -> MemoryDatabase -> Maybe (Dirty UnspentVal)
+getUnspentH op db = do
+    m <- M.lookup (outPointHash op) (hUnspent db)
+    I.lookup (fromIntegral (outPointIndex op)) m
+
+insertUnspentH :: Unspent -> MemoryDatabase -> MemoryDatabase
+insertUnspentH u db =
+    db
+        { hUnspent =
+              M.insertWith
+                  (<>)
+                  (outPointHash (unspentPoint u))
+                  (I.singleton
+                       (fromIntegral (outPointIndex (unspentPoint u)))
+                       (Modified (snd (unspentToVal u))))
+                  (hUnspent db)
+        }
+
+deleteUnspentH :: OutPoint -> MemoryDatabase -> MemoryDatabase
+deleteUnspentH op db =
+    db
+        { hUnspent =
+              M.insertWith
+                  (<>)
+                  (outPointHash op)
+                  (I.singleton (fromIntegral (outPointIndex op)) Deleted)
+                  (hUnspent db)
+        }