+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Do not add xpubs to cache if node not in sync.
 - Do not import transactions to cache if no xpubs in cache.
 - Refactor transaction importing logic extremely.
+- Refactor memory database and writer.
 
 ### Fixed
 - Fix BTC double-count transaction bug.